+++ cmd/root.go
@@ -22,9 +22,9 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/dremio/dremio-diagnostic-collector/cmd/root/collection"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/root/helpers"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
-	"github.com/dremio/dremio-diagnostic-collector/collection"
 	"github.com/dremio/dremio-diagnostic-collector/kubernetes"
 	"github.com/dremio/dremio-diagnostic-collector/pkg/versions"
 	"github.com/dremio/dremio-diagnostic-collector/ssh"
