+++ cmd/root/collection/capture_test.go
