+++ cmd/root/collection/testdata/kubernetes/cronjob.json
