+++ cmd/root/collection/testdata/kubernetes/endpoints.json
