+++ cmd/root/collection/testdata/kubernetes/job.json
