+++ cmd/root/collection/testdata/kubernetes/nodes.json
