+++ cmd/root/collection/testdata/kubernetes/replicaset.json
