+++ cmd/root/collection/testdata/kubernetes/sc.json
