+++ cmd/root/collection/testdata/kubernetes/service.json
