+++ cmd/root/collection/testdata/kubernetes/summary.json
